moved linkedIn functions to helpers folder and changed sendgridTemplate from var to let
