+++ backend/controllers/userController.js
@@ -1,6 +1,4 @@
 const _ = require('lodash');
-const axios = require('axios');
-const queryString = require('query-string');
 const util = require('../lib/utils');
 const config = require('../config/config');
 
@@ -11,37 +9,7 @@ const constants = require('../lib/constants');
 const Match = require('../models/match');
 const Token = require('../models/token');
 const User = require('../models/user');
-const { handleUserRegistration } = require('../helpers/user');
-
-const linkedinAuth = axios.create({
-  baseURL: 'https://www.linkedin.com/',
-});
-
-const linkedinApi = axios.create({
-  baseURL: 'https://api.linkedin.com/',
-});
-
-const extractImageUrls = (profileResponse) => {
-  const res = {};
-  if (profileResponse.data.profilePicture) {
-    const displayImages =
-      profileResponse.data.profilePicture['displayImage~'].elements;
-
-    displayImages.forEach((element) => {
-      const { displaySize } =
-        element.data['com.linkedin.digitalmedia.mediaartifact.StillImage'];
-      const imageKey = `${displaySize.width}x${displaySize.height}`;
-      const imageUrl = element.identifiers[0].identifier;
-
-      res[imageKey] = imageUrl;
-
-      if (imageKey == '200x200') {
-        res.default = imageUrl;
-      }
-    });
-  }
-  return res;
-};
+const { handleUserRegistration, getLinkedInProfile } = require('../helpers/user');
 
 const fetchUserToken = async (user) => {
   const token = await Token.findOne({ userId: user._id }).exec();
@@ -62,7 +30,7 @@ const sendRegistrationMail = async (user) => {
   const confirmationToken = util.encodeRegistrationToken(user._id);
   const confirmationLink = `https://${process.env.FRONTEND_BASE_URL}/confirmUserRegistration?confirmationToken=${confirmationToken}`;
 
-  var sendgridTemplate = ""
+  let sendgridTemplate = ""
   if(user.userType == constants.userTypes.mentee){
     sendgridTemplate = config.sendgrid.templates.mentee_signup;
   } else if(user.userType == constants.userTypes.mentor) {
@@ -83,67 +51,6 @@ const sendRegistrationMail = async (user) => {
   );
   return response;
 };
-
-const getLinkedInProfile = (authCode, isLocal = false) =>
-  new Promise((resolve, reject) => {
-    let redirectUri = '';
-    if (isLocal) {
-      redirectUri = process.env.LINKEDIN_REDIRECT_URI_LOCAL;
-    } else {
-      redirectUri = process.env.LINKEDIN_REDIRECT_URI;
-    }
-
-    const authUrl = queryString.stringifyUrl({
-      url: '/oauth/v2/accessToken',
-      query: {
-        grant_type: 'authorization_code',
-        code: authCode,
-        redirect_uri: redirectUri,
-        client_id: process.env.LINKEDIN_CLIENT_ID,
-        client_secret: process.env.LINKEDIN_CLIENT_SECRET,
-      },
-    });
-    const profileUrl =
-      '/v2/me?projection=(id,localizedFirstName,localizedLastName,profilePicture(displayImage~:playableStreams))';
-    const emaileUrl = `/v2/emailAddress?q=members&projection=(elements*(handle~))`;
-
-    const results = {};
-
-    linkedinAuth
-      .post(authUrl) // exchange auth code for access token
-      .then((response) => {
-        results.access_token = response.data.access_token;
-        return linkedinApi.get(profileUrl, {
-          // get user profile
-          headers: {
-            Authorization: `Bearer ${results.access_token}`,
-          },
-        });
-      })
-      .then((profileResponse) => {
-        results.profileResponse = profileResponse;
-        results.profileImageUrls = extractImageUrls(profileResponse);
-        return linkedinApi.get(emaileUrl, {
-          headers: {
-            Authorization: `Bearer ${results.access_token}`,
-          },
-        });
-      })
-      .then((emailResponse) => {
-        const response = {
-          firstName: results.profileResponse.data.localizedFirstName,
-          lastName: results.profileResponse.data.localizedLastName,
-          linkedInId: results.profileResponse.data.id,
-          email: emailResponse.data.elements[0]['handle~'].emailAddress,
-          profileImageUrls: results.profileImageUrls,
-        };
-        resolve(response);
-      })
-      .catch((error) => {
-        console.log(error);
-        reject(new Error({ msg: 'Error in getLinkedInProfile }))' }));
-      });
-  });
 
 const loginUser = async (req, res) => {
   const { body } = req;
@@ -213,101 +120,6 @@ const loginUser = async (req, res) => {
     return res.status(500).json({
       success: false,
       error: 'Unable to login user',
-    });
-  }
-};
-
-const registerUser = (req, res) => {
-  const { body } = req;
-
-  if (!body) {
-    return res
-      .status(400)
-      .json({ success: false, error: 'request body is empty' });
-  }
-
-  if (body.type == 'linkedInSignup') {
-    if (!body.authCode) {
-      return res
-        .status(400)
-        .json({ success: false, error: 'authCode missing in the body' });
-    }
-    const results = {};
-
-    const isLocal = body.isLocal == true;
-
-    getLinkedInProfile(body.authCode, isLocal)
-      .then((linkedInProfile) => {
-        results.linkedInProfile = linkedInProfile;
-        return User.findOne({ linkedInId: linkedInProfile.linkedInId }).exec();
-      })
-      .then((user) => {
-        let userObj = user;
-        if (user) {
-          userObj = Object.assign(user, results.linkedInProfile);
-        } else {
-          userObj = new User(results.linkedInProfile);
-        }
-        userObj.profileImageUrls = results.linkedInProfile.profileImageUrls;
-        return userObj.save();
-      })
-      .then((updatedUser) => {
-        results.updatedUser = updatedUser;
-        return Token.findOne({ userId: updatedUser._id });
-      })
-      .then((token) => {
-        // if token isn't in our DB, store
-        if (!token) {
-          // encrypt information
-          const t = util.refreshToken(results.updatedUser._id);
-
-          Token.create({
-            refreshToken: t,
-            userId: results.updatedUser._id,
-          });
-        }
-        // send the access token
-        const accessToken = util.accessToken(results.updatedUser._id);
-
-        return (
-          res
-            // .cookie('accessToken', accessToken, {
-            //   sameSite: 'none',
-            //   secure: true,
-            // })
-            .json({
-              success: true,
-              token: accessToken,
-              user: {
-                _id: results.updatedUser._id,
-                firstName: results.linkedInProfile.firstName,
-                lastName: results.linkedInProfile.lastName,
-                linkedInId: results.linkedInProfile.linkedInId,
-                email: results.linkedInProfile.email,
-              },
-            })
-        );
-      })
-      .catch((err) => {
-        console.log(err);
-        return res.status(500).json({
-          success: false,
-          error: 'Error in registering the user',
-        });
-      });
-  } else if (body.type == 'completeRegistration') {
-    if (!body.user) {
-      return res.status(400).json({
-        success: false,
-        error: 'request body does not have user object',
-      });
-    }
-
-    return {};
-  } else {
-    return res.status(400).json({
-      success: false,
-      error: 'invalid type',
     });
   }
 };
@@ -554,7 +366,6 @@ const confirmRegistration = async (req, res) => {
 
 module.exports = {
   loginUser,
-  registerUser,
   tempAuth,
   updateUser,
   userInfo,
